Replace Codec with StorageEncode/Decode for values stored in RocksDB
We now require that stored values implement the StorageEncode trait
to encode them together with their codec. When reading a value, the
value needs to implement StorageDecode. Alternatively, one can also
store raw bytes into RocksDB as it is needed for state values, for
example.
