+++ crates/storage-rocksdb/src/keys.rs
@@ -8,7 +8,10 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use bytes::{Buf, BufMut, BytesMut};
+use anyhow::anyhow;
+use bytes::{Buf, BufMut, Bytes, BytesMut};
+use bytestring::ByteString;
+use prost::encoding::encoded_len_varint;
 
 pub trait TableKey: Sized + Send + 'static {
     fn is_complete(&self) -> bool;
@@ -132,7 +135,7 @@ macro_rules! define_table_key {
             #[inline]
             fn serialize_to<B: bytes::BufMut>(&self, bytes: &mut B) {
                 $(
-                crate::codec::serialize(&self.$element, bytes);
+                $crate::keys::serialize(&self.$element, bytes);
                 )+
             }
 
@@ -141,7 +144,7 @@ macro_rules! define_table_key {
                 let mut this: Self = Default::default();
 
                 $(
-                    this.$element = crate::codec::deserialize(bytes)?;
+                    this.$element = $crate::keys::deserialize(bytes)?;
                 )+
 
                 return Ok(this);
@@ -151,7 +154,7 @@ macro_rules! define_table_key {
             fn serialized_length(&self) -> usize {
                 let mut serialized_length = 0;
                 $(
-                    serialized_length += crate::codec::Codec::serialized_length(&self.$element);
+                    serialized_length += $crate::keys::KeyCodec::serialized_length(&self.$element);
                 )+
                 serialized_length
             }
@@ -161,3 +164,238 @@ macro_rules! define_table_key {
 
 use crate::TableKind;
 pub(crate) use define_table_key;
+use restate_storage_api::deduplication_table::ProducerId;
+use restate_storage_api::StorageError;
+use restate_types::identifiers::InvocationUuid;
+
+pub(crate) trait KeyCodec: Sized {
+    fn encode<B: BufMut>(&self, target: &mut B);
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self>;
+
+    fn serialized_length(&self) -> usize;
+}
+
+impl KeyCodec for Bytes {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        write_delimited(self, target);
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        read_delimited(source)
+    }
+
+    fn serialized_length(&self) -> usize {
+        self.len()
+            + encoded_len_varint(u64::try_from(self.len()).expect("usize should fit into u64"))
+    }
+}
+
+impl KeyCodec for ByteString {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        write_delimited(self, target);
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        let bs = read_delimited(source)?;
+
+        unsafe { Ok(ByteString::from_bytes_unchecked(bs)) }
+    }
+
+    fn serialized_length(&self) -> usize {
+        self.len()
+            + encoded_len_varint(u64::try_from(self.len()).expect("usize should fit into u64"))
+    }
+}
+
+impl KeyCodec for u64 {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        target.put_u64(*self);
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        Ok(source.get_u64())
+    }
+
+    fn serialized_length(&self) -> usize {
+        8
+    }
+}
+
+impl KeyCodec for u32 {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        target.put_u32(*self);
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        Ok(source.get_u32())
+    }
+
+    fn serialized_length(&self) -> usize {
+        4
+    }
+}
+
+///
+/// Blanket implementation for Option.
+///
+impl<T: KeyCodec> KeyCodec for Option<T> {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        if let Some(t) = self {
+            t.encode(target);
+        }
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        if !source.has_remaining() {
+            return Ok(None);
+        }
+        let res = T::decode(source)?;
+        Ok(Some(res))
+    }
+
+    fn serialized_length(&self) -> usize {
+        self.as_ref().map(|v| v.serialized_length()).unwrap_or(0)
+    }
+}
+
+impl<'a> KeyCodec for &'a [u8] {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        target.put(*self);
+    }
+
+    fn decode<B: Buf>(_source: &mut B) -> crate::Result<Self> {
+        unimplemented!("could not decode into a slice u8");
+    }
+
+    fn serialized_length(&self) -> usize {
+        self.len()
+    }
+}
+
+impl KeyCodec for InvocationUuid {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        let slice = self.to_bytes();
+        debug_assert_eq!(slice.len(), self.serialized_length());
+        target.put_slice(&slice);
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        // note: this is a zero-copy when the source is bytes::Bytes.
+        if source.remaining() < InvocationUuid::SIZE_IN_BYTES {
+            return Err(StorageError::DataIntegrityError);
+        }
+        let bytes = source.copy_to_bytes(InvocationUuid::SIZE_IN_BYTES);
+        InvocationUuid::from_slice(&bytes).map_err(|err| StorageError::Generic(err.into()))
+    }
+
+    fn serialized_length(&self) -> usize {
+        InvocationUuid::SIZE_IN_BYTES
+    }
+}
+
+#[inline]
+fn write_delimited<B: BufMut>(source: impl AsRef<[u8]>, target: &mut B) {
+    let source = source.as_ref();
+    prost::encoding::encode_varint(source.len() as u64, target);
+    target.put(source);
+}
+
+#[inline]
+fn read_delimited<B: Buf>(source: &mut B) -> crate::Result<Bytes> {
+    let len = prost::encoding::decode_varint(source)
+        .map_err(|error| StorageError::Generic(error.into()))?;
+    // note: this is a zero-copy when the source is bytes::Bytes.
+    Ok(source.copy_to_bytes(len as usize))
+}
+
+#[inline]
+pub(crate) fn serialize<T: KeyCodec, B: BufMut>(what: &T, target: &mut B) {
+    what.encode(target);
+}
+
+#[inline]
+pub(crate) fn deserialize<T: KeyCodec, B: Buf>(source: &mut B) -> crate::Result<T> {
+    T::decode(source)
+}
+
+impl KeyCodec for ProducerId {
+    fn encode<B: BufMut>(&self, target: &mut B) {
+        match self {
+            ProducerId::Partition(p) => {
+                target.put_u8(0);
+                KeyCodec::encode(p, target)
+            }
+            ProducerId::Other(i) => {
+                target.put_u8(1);
+                KeyCodec::encode(i, target)
+            }
+        }
+    }
+
+    fn decode<B: Buf>(source: &mut B) -> crate::Result<Self> {
+        Ok(match source.get_u8() {
+            0 => ProducerId::Partition(KeyCodec::decode(source)?),
+            1 => ProducerId::Other(KeyCodec::decode(source)?),
+            i => {
+                return Err(StorageError::Generic(anyhow!(
+                    "Unexpected wrong discriminator for SequenceNumberSource: {}",
+                    i
+                )))
+            }
+        })
+    }
+
+    fn serialized_length(&self) -> usize {
+        1 + match self {
+            ProducerId::Partition(p) => KeyCodec::serialized_length(p),
+            ProducerId::Other(i) => KeyCodec::serialized_length(i),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use bytes::BytesMut;
+
+    #[test]
+    fn write_read_round_trip() {
+        let mut buf = BytesMut::new();
+        write_delimited("hello", &mut buf);
+        write_delimited(" ", &mut buf);
+        write_delimited("world", &mut buf);
+
+        let mut got = buf.freeze();
+        assert_eq!(read_delimited(&mut got).unwrap(), "hello");
+        assert_eq!(read_delimited(&mut got).unwrap(), " ");
+        assert_eq!(read_delimited(&mut got).unwrap(), "world");
+    }
+
+    fn concat(a: &'static str, b: &'static str) -> Bytes {
+        let mut buf = BytesMut::new();
+        write_delimited(a, &mut buf);
+        write_delimited(b, &mut buf);
+        buf.freeze()
+    }
+
+    #[test]
+    fn write_delim_keeps_lexicographical_sorting() {
+        assert!(concat("a", "b") < concat("a", "c"));
+        assert!(concat("a", "") < concat("d", ""));
+    }
+
+    #[test]
+    fn invocation_uuid_roundtrip() {
+        let uuid = InvocationUuid::new();
+
+        let mut buf = BytesMut::new();
+        uuid.encode(&mut buf);
+
+        let mut got_bytes = buf.freeze();
+
+        assert_eq!(got_bytes.len(), uuid.serialized_length());
+        let got = InvocationUuid::decode(&mut got_bytes).expect("deserialization should work");
+
+        assert_eq!(uuid, got);
+    }
+}
